bazar_marketplace/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
alembic
pydantic
python-dotenv